// src/app/models/child.ts
export class Child {
    id: number;
    firstName: string;
    lastName: string;
    birthDate: Date;
    street: string;
    city: string;
    province: string;
    postalCode: string;
    country: string;
    latitude: number;
    longitude: number;
    isNaughty: boolean;
    dateCreated: Date;
}